requirements.txt
numpy==2.4.4
PyYAML==6.0.3
Requests==2.33.1
websockets==16.0
